+++ packages/cli/src/index.ts
@@ -11,7 +11,10 @@ program
   .command('create-app', 'Create an app from one of our examples', {
     executableFile: resolve(__dirname, '../dist/create-app.js')
   })
-  .command('cms-sync-s3', 'Sync content files from a S3 to the file system', {
+  .command('cms-sync-s3', 'Sync content files from S3 to the file system', {
     executableFile: resolve(__dirname, '../dist/cms-sync-s3.js')
+  })
+  .command('init-s3-sync', 'Initialize the S3 cache and create the contentful webhook for it', {
+    executableFile: resolve(__dirname, '../dist/init-s3-sync.js')
   })
   .parse(process.argv);