+++ packages/contentful-fs-loader/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @last-rev/contentful-fs-loader
 
+## 0.1.5
+
+### Patch Changes
+
+- 63176af: Added logging and timers, added a new cms data loader
+
+## 0.1.4
+
+### Patch Changes
+
+- a09b949: Removed env vars from gql server, requiring all to be passed in
+
+## 0.1.3
+
+### Patch Changes
+
+- ccc1937: Changed the FS structure to support atomic files for lookups and contentTypes
+
 ## 0.1.2
 
 ### Patch Changes