+++ packages/graphql-contentful-core/CHANGELOG.md
@@ -1,5 +1,48 @@
 # @last-rev/graphql-contentful-core
 
+## 0.1.8
+
+### Patch Changes
+
+- 63176af: Added logging and timers, added a new cms data loader
+- Updated dependencies [63176af]
+  - @last-rev/contentful-fs-loader@0.1.5
+  - @last-rev/contentful-s3-loader@0.1.3
+  - @last-rev/graphql-schema-gen@0.1.5
+
+## 0.1.7
+
+### Patch Changes
+
+- 94b5414: Fixed gql handler, and issues with s3 loader
+- Updated dependencies [94b5414]
+  - @last-rev/contentful-s3-loader@0.1.2
+
+## 0.1.6
+
+### Patch Changes
+
+- 72044a3: Updated sschema gen to pass an array for theme
+- Updated dependencies [72044a3]
+  - @last-rev/graphql-schema-gen@0.1.4
+
+## 0.1.5
+
+### Patch Changes
+
+- a09b949: Removed env vars from gql server, requiring all to be passed in
+- Updated dependencies [a09b949]
+  - @last-rev/contentful-fs-loader@0.1.4
+  - @last-rev/contentful-s3-loader@0.1.1
+
+## 0.1.4
+
+### Patch Changes
+
+- ccc1937: Changed the FS structure to support atomic files for lookups and contentTypes
+- Updated dependencies [ccc1937]
+  - @last-rev/contentful-fs-loader@0.1.3
+
 ## 0.1.3
 
 ### Patch Changes