+++ packages/graphql-schema-gen/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @last-rev/graphql-schema-gen
 
+## 0.1.5
+
+### Patch Changes
+
+- 63176af: Added logging and timers, added a new cms data loader
+
+## 0.1.4
+
+### Patch Changes
+
+- 72044a3: Updated sschema gen to pass an array for theme
+
 ## 0.1.3
 
 ### Patch Changes